### 2021-09-03 #3688 @whedon
[PRE REVIEW]: electiondata: a Python package for consolidating, checking, analyzing, visualizing and exporting election results
**Submitting author:** @sfsinger19103 (<a href="http://orcid.org/0000-0002-1373-9619">Stephanie Singer</a>)
**Repository:** <a href="https://github.com/ElectionDataAnalysis/electiondata" target ="_blank">https://github.com/ElectionDataAnalysis/electiondata</a>
**Version:** v2.0.0
