### 2021-09-16 df0b1eb
Redo naming and documentation to indicate that the NIST json format is v2, not v1
