+++ README.md
@@ -2,22 +2,28 @@
 
 
 # Overview
-This repository hopes to provide reliable tools for consolidation and analysis of raw election results from the most reliable sources -- the election agencies themselves. 
+This repository provides tools for consolidation and analysis of raw election results from the most reliable sources -- the election agencies themselves. 
  * Consolidation: take as input election results files from a wide variety of sources and load the data into a relational database
- * Export: create tab-separated flat export files of results sets rolled up to any desired intermediate geography (e.g., by county, or by congressional district)
- * Analysis: provide a variety of analysis tools
- * Visualization: provide a variety of visualization tools.
+ * Export: create consistent-format export files of results sets rolled up to any desired intermediate geography
+   * tabular (tab-separated text)
+   * xml (following NIST Election Results Reporting Common Data Format V2)
+   * json (following NIST Election Results Reporting Common Data Format V2)
+ * Analysis: 
+   * Curates one-county outliers of interest
+   * Calculates difference-in-difference for reaults available by vote type
+ * Visualization: 
+   * Scatter plots
+   * Bar charts
 
 # Target Audience
 This system is intended to be of use to candidates and campaigns, election officials, students of politics and elections, and anyone else who is interested in assembling and understanding election results.
 
 # How to Contribute Code
-Please contribute code that works in python 3.7, with the package versions specified in [requirements.txt](requirements.txt). We follow the [black](https://pypi.org/project/black/) format.
+Please contribute code that works in python 3.9, with the package versions specified in [requirements.txt](requirements.txt). We follow the [black](https://pypi.org/project/black/) format.
 
 # How to Help in Other Ways
 If you have skills to contribute to building the system, we can definitely use your help:
  * Creating visualizations
- * Importing and exporting data via xml feeds
  * Preparing for intake of specific states' results files
  * Managing collection of data files in real time
  * Writing documentation
@@ -25,7 +31,7 @@ If you have skills to contribute to building the system, we can definitely use y
  * Building our open source community
  * What else? Let us know!
  
-If you are a potential end user -- an election official, political scientist or campaign consultant, for instance -- we would love to talk with you about what you want to from this system.
+If you are a potential end user -- an election official, political scientist or campaign consultant, for instance -- let us know what you want to from this system.
  
 If you are interested in contributing, or just staying updated on the progress of this project, please [contact Stephanie Singer](http://symmetrysinger.com/index.php?id=contact). 
 
@@ -45,6 +51,7 @@ Detailed instructions can be found [here](docs/User_Guide.md).
 Funding provided October 2019 - September 2021 by the National Science Foundation
  * Award #1936809, "EAGER: Data Science for Election Verification" 
  * Award #2027089, "RAPID: Election Result Anomaly Detection for 2020"
+Data collection and consolidation for the 2020 US General Election funded in part by the Verified Voting Foundation.
 
 # License
 See [LICENSE.md](./LICENSE.md)
