### 2021-09-20 #3739 @sfsinger19103
Paper is (and has been) here: https://github.com/ElectionDataAnalysis/electiondata/blob/joss-submission/JoSS_Submission/paper.md

Not sure why @whedon couldn't find it. Please let me know what to do to fix the problem.
