### 2021-09-21 3c39c24
rename test function; remove line break in nh20g_CD2_official.ini
