### 2021-09-21 3305602
concatenate list of dfs once rather than iteratively
