### 2021-09-21 dc64d20
better error messaging
