### 2021-09-22 548c724
Kansas munging improvements
