### 2021-10-04 #3739 @vaneseltine
Currently getting prerequisites in place (PostgreSQL, Python 3.9) Two comments while I run through setup:

1. There's a bit of inconsistency in Python version. The instructions in User_Guide.md require Python 3.9, but the main README.md only mentions 3.7 for contributions, and no version requirement is enforced (e.g. with python_requires) in setup.py.

2. This requirement in the user guide is best to avoid: "If you use the alias python, make sure it points to python3.9." In most cases, in a Python-based project, calling literal "python" can be entirely avoided. And doing a couple of searches through the code, I'm not sure why this seems important for electiondata. Happy to help troubleshoot if there's a point in the code where this is an issue.
