### 2021-10-04 #3739 @sfsinger19103
1. Thanks for flagging the version inconsistency. It should be 3.9
2. The comment about aliasing is not intended to be a requirement, but rather as a helpful hint to users who may not have much experience with python.
