### 2021-11-04 #3739 @andrewheiss
Hi! Sorry for the delay on this! Here's my review:

---

**electiondata** is a powerful new Python package that allows users to read and standardize raw election data that state and local election agencies provide in a wide variety of file formats and data structures. It is important software for anyone interested in election data, which is ordinarily incredibly messy and hard to work with, and will make it much easier to analyze and visualize election results in a standard way.

## Installation

In the paper, the authors state that “Users do not need to know python (other than the basics for installing and calling the package)”, but I found the installation process more difficult than expected. Using a clean, new macOS Monterrey installation with freshly installed Python 3.9, without using any virtual environments or conda, and following the installation instructions in the user guide, I ran into these surprises:

- I ran `python3 setup.py install`, following the instructions, and it seemed that the package and all dependencies installed correctly. But running `import electiondata as ea` from the Python shell gave errors about missing various dependencies (`numpy`, `lxml`, and others).
- So I ran `pip3 install -r requirements.txt` to install from the requirements file instead, and that mostly worked, except…
	- …it’s not immediately clear in the installation section of the guide that PostgreSQL is required (and is specifically a prerequisite for installing the package; `psycopg2` wouldn’t install through `pip` without it on my system). Running `brew install postgresql` fixed this just fine.
	- `matplotlib` wouldn’t install because I was missing freetype. `brew install freetype` fixed this.
	- After installing PostgreSQL and Freetype, `pip3 install -r requirements.txt` worked fine and the package installed; running `import electiondata as ea` in the Python shell produced no errors

These ↑ are all issues specific to a clean installation of macOS, so it’s not anything that can be universally addressed (and Windows and Linux will have their own issues), but it might be helpful to expand the installation section of the user guide to specify that external programs like PostgreSQL and Freetype are necessary before installing `electiondata`.

Additionally, it might be helpful to include basic instructions for how to create a new local PostgreSQL user and database so those settings and credentials can be included in `run_time.ini`. If the package’s goal is to be accessible to people with minimal Python experience, providing this extra guidance would be helpful (it’s been a few years since I last used PostgreSQL, for instance, so I had to google how to create a new user and db).

It’s also not clear in the user guide what needs to happen to set up the initial database. Before trying to load any data, I poked around in the package code to find what the expected schema should be, expecting to need to create different tables on my own, but then discovered that the requisite tables are built by the `create_database()` function in `src/electiondata/database/__init__.py` upon importing the package, which was a nice surprise. Documenting that process could be helpful.

Finally (and this is definitely more of a distant feature request/idea!), it might be worth it to someday use something more local like SQLite by default, which lets you embed the database directly into the application and doesn’t require a separate SQL server process. It would simplify code testing and development, since users wouldn’t need to have a separate server running. Currently it is possible to do something like SQLite—the documentation points out that users can modify `src/database/__init__.py` to use other flavors of SQL, and the package uses `squalchemy` to provide a standardized API for all sorts of SQL backends, including SQLite—but switching between PostgreSQL and SQLite is nontrivial. Again, this is definitely not a dealbreaker at all! Just a future thought to simplify the installation and onboarding for the package.

## Documentation and onboarding

The user guide is comprehensive and works well as a complete documentation reference, but it is very difficult to get started initially. There are sample .ini files, but they’re scattered throughout the project.

It would be really helpful to have a sort of “Getting Started” vignette or guide that showed a minimal working example of how to:

1. Structure the project directory with the different `[electiondata]` folders like `results_dir`, `repository_content_dir`, etc.
2. Generate the jurisdiction files for a known state (like Georgia, since some raw election data is already included in the tests)
3. Load the data from one of the raw election datasets (like Georgia’s XML file or Guam’s Excel file from the tests)
4. Analyze the loaded data (again, for Georgia or Guam, since those raw datasets are included with the package already)

Again, the current user guide is quite comprehensive and covers all sorts of edge cases, conventions, additional arguments, and other details, but it’s hard to use it to get started from scratch. (I had to figure out lots of the program flow from the tests, but I eventually got it to work and load Georgia and Guam data into my local PostgreSQL database)

A complete working example that users can run would help them understand the flow of the package and ensure that all the moving parts are working before they start feeding their own downloaded data files into the database.

## Testing

There is a suite of tests for each of the package’s main purposes: jurisdiction creation, data loading, and analyzing. These run well and are sufficient.

## Contributions

The README includes a section on contributing to the project, specifying that code should be compatible with Python 3.7 (though this should be 3.9), and that it should follow black linting style.

It might be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md)), in order to meet JOSS’s requirements that guidelines tell people how to (1) contribute, (2) report issues, and (3) seek support.

---

In general this is great! My only big concerns really just deal with getting it all set up and working and running—a vignette with a minimal working reproducible example that takes a raw XML/Excel/JSON/whatever file from its raw state into the database into nice graphs would be extraordinarily helpful and would make it a lot easier to help people start using this package right away.
