### 2021-11-04 #3739 @ajstewartlang
Many thanks for your very detailed and helpful review @andrewheiss

@sfsinger19103 would you be able to address the points that are raised? On my reading, it does sound like additional installation instructions would be helpful to the novice Python user. Adding a 'Getting Started' vignette sounds like another great suggestion. The other points raised are important too - espec. in terms of adding/improving documentation and giving a minimal working reproducible example. This looks like a great submission and addressing these points will likely make the package even more usable and likely increase adoption. The helpful suggestions raised by @andrewheiss sound very do-able to me so I hope you can address them. @sfsinger19103 it might be easiest if you open a separate issue for each so we can then tick each of them off as they are addressed. Does this sound ok?
