### 2021-11-04 ElectionDataAnalysis/electiondata#727 @sfsinger19103
Create issues for items in JOSS review
See Nov 3 review in [this thread](https://github.com/openjournals/joss-reviews/issues/3739). Create separate issues here for the suggestions there.
