### 2021-11-04 ElectionDataAnalysis/electiondata#731 @sfsinger19103
Community guidelines
It might be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md)), in order to meet JOSS’s requirements that guidelines tell people how to (1) contribute, (2) report issues, and (3) seek support.
