### 2021-11-04 #3739 @sfsinger19103
@ajstewartlang what's the appropriate way to make a small correction to the submitted paper?

(Specifically, the paper says that there is an export in json format that conforms to the NIST Common Data Format V1 when in fact the json format conforms to NIST Common Data Format V2.)
