### 2021-11-05 b61f8cd
code pre-process for NJ (WIP)
