### 2021-11-05 13f043c
Expand installation instructions
