+++ docs/User_Guide.md
@@ -3,10 +3,11 @@
 You will need:
  * `python3.9`. 
  * `postgresql`. You do *not* need to create a database; you just need to make sure a `postgresql` is installed, and you need to have the hostname, port and login credentials with default permissions. 
+ * all required python packages: run `python3.9 -m pip install -r requirements.txt` from the root folder of the repository.
    
 For more detail, see the [Installation Instructions](./install.md).
 
-To use other varieties of SQL, you would need to modify the routines in the `database` module. 
+To use other varieties of SQL, you would need to modify the routines in the `database` module.
 
 ## Installation
 The one-line version: From the root folder of your repository run `python3 setup.py install`. For more detail, see the [Installation Instructions](./install.md).
