### 2021-11-05 #3739 @ajstewartlang
@sfsinger19103 you can go ahead and correct the paper on the joss-submission branch (or on the joss-submission-updated branch if that's where changes have been made) - I'll then re-generate the pdf - just let me know which branch it's on.
