### 2021-11-07 503dddc
Create sample session document (WIP)
