### 2021-11-07 897bd57
Create sample session document (WIP)
