### 2021-11-09 1979a46
Create sample session document (WIP)
