### 2021-11-09 f2f467a
Create sample session document (WIP)
