### 2021-11-13 31594b6
Create sample session document (WIP)
