+++ docs/User_Guide.md
@@ -157,7 +157,7 @@ Sometimes it is necessary to use regular expressions to extract information from
 
 The system will report (in the `.warnings` files) any strings that did not match the regex. 
 
-For an interactive sandbox for learning python regex, try `RegularExpression101.com`.
+For an interactive sandbox for learning python regex, try `regex101.com`.
 
 ### \[\<foreign key\> lookup\]
 
@@ -231,16 +231,17 @@ In addition, `new_juris_files` creates a starter dictionary `XX_starter_dictiona
 
 NB: it is perfectly OK to have more than one raw_identifier_value for a single element. This can be necessary if, say, different counties use different names for a single contest. What can cause problems are lines with the same cdf_element and same raw_identifier_value, but different cdf_internal_names.
 
- (3) Add all counties to the `ReportingUnit.txt` file and `XX_starter_dictionary.txt`. You must obey the semicolon convention so that the system will know that the counties are subunits of the jurisdiction. For example:
+ (3) Add all counties to the `ReportingUnit.txt` file and `XX_starter_dictionary.txt`. 
+ 
+The `ReportingUnit.txt` file determines the names used internally in your database. (We have found that including "County" in the name helps with disambiguation.) You must obey the semicolon convention so that the system will know that the counties are subunits of the jurisdiction. For example:
 ```
 Name	ReportingUnitType
 Texas;Angelina County	county
 Texas;Gregg County	county
 Texas;Harrison County	county
 ```
-Counties must be added by hand. 
 
-NB: in some jurisdictions, the major subdivision type is not 'county. For instance, Louisiana's major subdivisions are called 'parish'. In the `elections.analyze` module, several routines roll up results to the major subdivision -- usually counties. By default, the ReportingUnitType of the major subdivision is read from the file [major_subjurisdiction_types.txt](../src/jurisdictions/000_for_all_jurisdictions/major_subjurisdiction_types.txt) if possible; if that file is missing, or does not provide a subdivision type for the particular jurisdiction in question, the system will try to deduce the major subdivision type from the database. A different file of subdivision types can be specified with the optional `major_subdivision_file` parameter in `Analyzer()` or `DataLoader()`
+NB: in some jurisdictions, the major subdivision type is not 'county'. For instance, Louisiana's major subdivisions are called 'parish'. In the `elections.analyze` module, several routines roll up results to the major subdivision -- usually counties. By default, the ReportingUnitType of the major subdivision is read from the file [major_subjurisdiction_types.txt](../src/jurisdictions/000_for_all_jurisdictions/major_subjurisdiction_types.txt) if possible; if that file is missing, or does not provide a subdivision type for the particular jurisdiction in question, the system will try to deduce the major subdivision type from the database. A different file of subdivision types can be specified with the optional `major_subdivision_file` parameter in `Analyzer()` or `DataLoader()`
 
 The system assumes that internal database names of ReportingUnits carry information about the nesting of the basic ReportingUnits (e.g., counties, towns, wards, etc., but not congressional districts) via semicolons. For example: `
  * `Pennsylvania;Philadelphia;Ward 8;Division 6` is a precinct in 
@@ -250,7 +251,7 @@ The system assumes that internal database names of ReportingUnits carry informat
  
 Other nesting relationships (e.g., `Pennsylvania;Philadelphia;Ward 8;Division 6` is in `Pennsylvania;PA Senate District 1`) are not yet recorded in the system (as of 4/2/2021).
 
-To find the raw_identifiers for the dictionary, look in your results files to see how counties are written. For example, if your results file looks like this (example from Texas):
[…]
