+++ src/jurisdictions/Georgia/CandidateContest.txt
@@ -1,0 +1,258 @@
+Name	NumberElected	Office	PrimaryParty
+GA House District 1	1	GA House District 1	
+GA House District 2	1	GA House District 2	
+GA House District 3	1	GA House District 3	
+GA House District 4	1	GA House District 4	
+GA House District 5	1	GA House District 5	
+GA House District 6	1	GA House District 6	
+GA House District 7	1	GA House District 7	
+GA House District 8	1	GA House District 8	
+GA House District 9	1	GA House District 9	
+GA House District 10	1	GA House District 10	
+GA House District 11	1	GA House District 11	
+GA House District 12	1	GA House District 12	
+GA House District 13	1	GA House District 13	
+GA House District 14	1	GA House District 14	
+GA House District 15	1	GA House District 15	
+GA House District 16	1	GA House District 16	
+GA House District 17	1	GA House District 17	
+GA House District 18	1	GA House District 18	
+GA House District 19	1	GA House District 19	
+GA House District 20	1	GA House District 20	
+GA House District 21	1	GA House District 21	
+GA House District 22	1	GA House District 22	
+GA House District 23	1	GA House District 23	
+GA House District 24	1	GA House District 24	
+GA House District 25	1	GA House District 25	
+GA House District 26	1	GA House District 26	
+GA House District 27	1	GA House District 27	
+GA House District 28	1	GA House District 28	
+GA House District 29	1	GA House District 29	
+GA House District 30	1	GA House District 30	
+GA House District 31	1	GA House District 31	
+GA House District 32	1	GA House District 32	
+GA House District 33	1	GA House District 33	
+GA House District 34	1	GA House District 34	
+GA House District 35	1	GA House District 35	
+GA House District 36	1	GA House District 36	
+GA House District 37	1	GA House District 37	
+GA House District 38	1	GA House District 38	
+GA House District 39	1	GA House District 39	
+GA House District 40	1	GA House District 40	
+GA House District 41	1	GA House District 41	
+GA House District 42	1	GA House District 42	
+GA House District 43	1	GA House District 43	
+GA House District 44	1	GA House District 44	
+GA House District 45	1	GA House District 45	
+GA House District 46	1	GA House District 46	
+GA House District 47	1	GA House District 47	
+GA House District 48	1	GA House District 48	
+GA House District 49	1	GA House District 49	
+GA House District 50	1	GA House District 50	
+GA House District 51	1	GA House District 51	
+GA House District 52	1	GA House District 52	
+GA House District 53	1	GA House District 53	
+GA House District 54	1	GA House District 54	
+GA House District 55	1	GA House District 55	
+GA House District 56	1	GA House District 56	
+GA House District 57	1	GA House District 57	
+GA House District 58	1	GA House District 58	
+GA House District 59	1	GA House District 59	
+GA House District 60	1	GA House District 60	
+GA House District 61	1	GA House District 61	
+GA House District 62	1	GA House District 62	
+GA House District 63	1	GA House District 63	
+GA House District 64	1	GA House District 64	
+GA House District 65	1	GA House District 65	
+GA House District 66	1	GA House District 66	
+GA House District 67	1	GA House District 67	
+GA House District 68	1	GA House District 68	
+GA House District 69	1	GA House District 69	
+GA House District 70	1	GA House District 70	
+GA House District 71	1	GA House District 71	
+GA House District 72	1	GA House District 72	
+GA House District 73	1	GA House District 73	
+GA House District 74	1	GA House District 74	
+GA House District 75	1	GA House District 75	
+GA House District 76	1	GA House District 76	
[…]

[…]
