### 2021-11-13 d84b6a6
Create sample session document (WIP)
