+++ src/jurisdictions/Georgia/Party.txt
@@ -1,5 +1,6 @@
 Name
 Democratic Party
 Republican Party
-Green Party
 Libertarian Party
+Green Party
+Independent Party
