### 2021-11-23 #3739 @vaneseltine
Apologies for my delays, but I've now been able to spend a good chunk of time on this.

---

### Fix: Community guidelines

As @andrewheiss mentioned, a little more guidance for potential contributors or those with problems would be helpful. I see you have an issue up https://github.com/ElectionDataAnalysis/electiondata/issues/731 -- this doesn't need to be extensive, but it's good to have some clear directions for those who land on the GitHub page looking for them.

### Suggestion: Improvements to User Guide

I don't see this as a requirement, but please take it as input for updates or revisions that you might choose to do with the User Guide. I largely agree with what @andrewheiss wrote about a minimum reproducible example and the difficulty of getting up and running.

In particular, it took me a while to understand what `results_dir` in `run_time.ini` needed to contain (not, e.g., `reference_results`). Ultimately I figured out -- by reading through the tests -- that I could download https://github.com/ElectionDataAnalysis/TestingData and make use of that as a results base. Once I saw how that was structured, I could also/instead use a copy of `electiondata/tests/000_data_for_pytest/2020-General`.

For me at least, it would have been very helpful for the User Guide to provide something like an overview of the entire package and its major processes along with more concrete example(s). In hindsight, `</path/to/folder/containing/result/files/for/batch/upload>` does indeed make sense, but I would have sorted this out much more easily if the User Guide provided, for example, a screenshot or ASCII representation of an example folder layout and workflow. I got there in the end, but the User Guide really dives into the minutia. Even a brief, summary overview of each of the key processes and how they all fit together would be, I think, a great help.

### Fix: Handle file encoding

Having worked my way through some of the User Guide examples of results processing/handling with TestingData. I ran into a couple problems, particularly around Windows use, that I do think need fixing. First:

Using Windows, I ran into a UnicodeEncodeError from load_all() on the TestingData repo -- which fully crashed the process:

```
Munger warnings written to ./output\load_all_1123_1122\Alaska_munger_ak_gen_fwab.munger.warnings
Traceback (most recent call last):
  File "[...]\electiontest\testing.py", line 3, in <module>
    dl.load_all()
  File "[...]\electiondata\.venv\lib\site-packages\electiondata-2.0-py3.9.egg\electiondata\__init__.py", line 874, in load_all
    err = ui.report(
    f.write(out_str)
  File "C:\Python\39\lib\encodings\cp1252.py", line 19, in encode
    return codecs.charmap_encode(input,self.errors,encoding_table)[0]
UnicodeEncodeError: 'charmap' codec can't encode character '\x96' in position 969: character maps to <undefined>
```

I'm assuming it was just trying to write another warning, as I could re-run to resume the processing without further error. My assumption is that one of these input files was UTF-8 encoded, and could not be interpreted as Window's default CP-1252.

You should ensure that the program is Unicode-aware. For file handling see, e.g., https://stackoverflow.com/questions/19591458/python-reading-from-a-file-and-saving-to-utf-8 and for the full story, https://docs.python.org/3/howto/unicode.html

### Fix: Convert or "slugify" file names

Second, Windows/NTSF filenames have somewhat more restrictive rules than Linux/UNIX-based systems, and so (for example) this attempt to provide warnings based on information with multiple colons was not possible:

```
Traceback (most recent call last):
  File "[...]\electiontest\testing.py", line 3, in <module>
    dl.load_all()
  File "[...]\electiondata\.venv\lib\site-packages\electiondata-2.0-py3.9.egg\electiondata\__init__.py", line 874, in load_all
    err = ui.report(
  File "[...]\electiondata\.venv\lib\site-packages\electiondata-2.0-py3.9.egg\electiondata\userinterface\__init__.py", line 728, in report
    with open(out_path, "a") as f:
OSError: [Errno 22] Invalid argument: './output\\load_all_1123_1355\\Maricopa-County_test_Election: 2020 Primary; Jurisdiction: Maricopa County.warnings'
```

There are a bunch of ways to handle the problem of turning arbitrary strings into valid filenames, see e.g., https://stackoverflow.com/questions/295135/turn-a-string-into-a-valid-filename. I'd suggest slugify() from https://github.com/un33k/python-slugify (via pip as python-slugify) for simplicity and ease of adjustment as desired:

```
>>> from slugify import slugify
>>> slugify("Maricopa-County_test_Election: 2020 Primary; Jurisdiction: Maricopa County")
'maricopa-county-test-election-2020-primary-jurisdiction-maricopa-county'
>>> slugify("Maricopa-County_test_Election: 2020 Primary; Jurisdiction: Maricopa County", regex_pattern=r'[^ A-z0-9-_]+', lowercase=False)
'Maricopa-County_test_Election- 2020 Primary- Jurisdiction- Maricopa County'
```

### Suggestion: Examine automated tests

I've uploaded logs of my pytest runs (again, on Windows) here: https://gist.github.com/vaneseltine/d493a9741d8319b6bec7ddc94b19e112

1. I'm not sure what the issue with dataloading_tests is -- it fails 1 test on an assertion about receiving the correct files.
2. analyzer_tests looks like it's getting tripped up on exact float requirements, and so fails 3 tests on the score due to inequalities like 2.3758572130853186 != 2.375857213085318
3. No problems with jurisdiction_prepper_tests

Nothing too alarming, but I wanted to provide you these logs for reference. #2 is probably another issue that boils down to a difference in platform/OS but it's worth considering a test strategy that will make a "close enough" comparison. pytest.approx() is designed to do this, but it can't work on large mixed-type dicts like the ones being compared there, so it's not a drop-in solution for you.

---

This is an impressive package that will really benefit from more first-time user guidance in the documentation and minor code revisions to ensure smooth operation across platforms.
