### 2021-11-28 a9c4b94
Create sample session document (WIP)
