### 2021-11-29 227a5d6
Create sample session document (WIP)
