### 2021-11-30 82f6052
Create sample session document (WIP)
