+++ docs/Sample_Session.md
@@ -6,7 +6,6 @@ This document walks the reader through a simple example, from setting up project
 The package offers a fair amount of flexibility in the directory structures used. For this sample session, we assume the user will call the program from a working directory with the following structure and files:
 ```
 .
-+-- jurisdiction_prep.ini
 +-- run_time.ini
 +-- results_directory
 |   +-- Georgia
@@ -34,25 +33,96 @@ password=
 ```
 You may wish to check that these postgresql credentials will work on your system via the command `psql -h localhost -p 5432 -U postgres postgres`. If this command fails, or if it prompts you for a password, you will need to find the correct connection parameters specific to your postgresql instance.  (Note that the `dbname` parameter is arbitrary, and determines only the name of the postgresql database created by the package.)
 
-### Contents of `jurisdiction_prep.ini` (optional)
-If you choose to create your Georgia jurisdiction files from scratch (rather than using the ones provided in the repository) you will need to specify various information about Georgia in `jurisdiction_prep.ini`:
-```
-[electiondata]
-name=Georgia
-reporting_unit_type=state
-abbreviated_name=GA
-count_of_state_house_districts=180
-count_of_state_senate_districts=56
-count_of_us_house_districts=14
-```
-
 ### Contents of `GA_detail_20201120_1237.xml`
 Copy the file of the same name in the repository: 000_template.mungerGA_detail_20201120_1237.xml](../tests/000_data_for_pytest/2020-General/Georgia/GA_detail_20201120_1237.xml)
 
-## Specify contests totals for data quality check (optional)
+## Create Georgia jurisdiction files (optional)
+If you wish to practice creating jurisdiction files, follow the steps below. Otherwise, skip to the next section.
+
+## Load Data
+```
+>>> import electiondata as ed
+>>> ed.load_or_reload_all()
+```
+After this command executes successfully, you will see a database in your postgres instance whose name matches the value of `dbname` given in `run_time.ini`. The `Georgia` results folder will be moved to the archive directory, with a date stamp from the date of the results file. There will be a summary of the results of the comparison with the reference results in the `reports_and_plots_dir`.
+```
+.
++-- run_time.ini
++-- results_directory
++-- archive_directory
+|   +-- Georgia
+|   |   +-- GA_detail_20201120_1237.xml
++-- reports_and_plots_directory
+|   +-- compare_to_Georgia_xxxx_xxxx
+|   |   +-- parameters.txt
+|   |   +-- not_found_in_db.tsv
+|   |   +-- ok.tsv
+|   |   +-- wrong.tsv
+```
+ 
+
+## Reading data from the database
+To pull data out, you will need to use the Analyzer class:
+```
+>>> an = ed.Analyzer()
+```
+
+## Export results
+You can export results in tabular form:
+```
+>>> an.export_election_to_tsv("GA_results.tsv", "2020 General")
+```
+The  file `GA_results.txv` containing the tab-separated results will be created in the working directory:
+```
+.
++-- archive_directory
+|   +-- Georgia
+|   |   +-- GA_detail_20201120_1237.xml
+|   GA_results.tsv 
++-- reports_and_plots_directory
+|   +-- compare_to_Georgia_xxxx_xxxx
+|   |   +-- parameters.txt
+|   |   +-- not_found_in_db.tsv
+|   |   +-- ok.tsv
+|   |   +-- wrong.tsv
++-- results_directory
++-- run_time.ini
+```
+
+The program can also produce a string of data in the NIST Common Data Format Version 2.0, in either json or xml format:
+```
+>>> results_string_xml = an.export_nist_xml_as_string("2020 General", "Georgia")
+>>> results_string_json = an.export_nist_json_as_string("2020 General", "Georgia")
+```
+
+## Analysis and Plots
[…]
