### 2021-11-30 78df391
Create sample session document (WIP)
