### 2021-11-30 e2917fb
Create sample session document (WIP)
