### 2021-11-30 7e1d489
Create sample session document (WIP)
