### 2021-11-30 356150f
Create sample session document (WIP)
