### 2021-11-30 01e8421
Create sample session document (WIP)
