### 2021-11-30 7e87036
Create sample session document (WIP)
