+++ src/jurisdictions/Georgia/BallotMeasureContest.txt
@@ -1,2 +1,1 @@
 Name	ElectionDistrict	Election
-Constitutional Amendment #1	Georgia	2020 General
