+++ src/jurisdictions/Georgia/ReportingUnit.txt
@@ -1,241 +1,10 @@
 Name	ReportingUnitType
-Georgia;GA House District 1	state-house
-Georgia;GA House District 2	state-house
-Georgia;GA House District 3	state-house
-Georgia;GA House District 4	state-house
-Georgia;GA House District 5	state-house
-Georgia;GA House District 6	state-house
-Georgia;GA House District 7	state-house
-Georgia;GA House District 8	state-house
-Georgia;GA House District 9	state-house
-Georgia;GA House District 10	state-house
-Georgia;GA House District 11	state-house
-Georgia;GA House District 12	state-house
-Georgia;GA House District 13	state-house
-Georgia;GA House District 14	state-house
-Georgia;GA House District 15	state-house
-Georgia;GA House District 16	state-house
-Georgia;GA House District 17	state-house
-Georgia;GA House District 18	state-house
-Georgia;GA House District 19	state-house
-Georgia;GA House District 20	state-house
-Georgia;GA House District 21	state-house
-Georgia;GA House District 22	state-house
-Georgia;GA House District 23	state-house
-Georgia;GA House District 24	state-house
-Georgia;GA House District 25	state-house
-Georgia;GA House District 26	state-house
-Georgia;GA House District 27	state-house
-Georgia;GA House District 28	state-house
-Georgia;GA House District 29	state-house
-Georgia;GA House District 30	state-house
-Georgia;GA House District 31	state-house
-Georgia;GA House District 32	state-house
-Georgia;GA House District 33	state-house
-Georgia;GA House District 34	state-house
-Georgia;GA House District 35	state-house
-Georgia;GA House District 36	state-house
-Georgia;GA House District 37	state-house
-Georgia;GA House District 38	state-house
-Georgia;GA House District 39	state-house
-Georgia;GA House District 40	state-house
-Georgia;GA House District 41	state-house
-Georgia;GA House District 42	state-house
-Georgia;GA House District 43	state-house
-Georgia;GA House District 44	state-house
-Georgia;GA House District 45	state-house
-Georgia;GA House District 46	state-house
-Georgia;GA House District 47	state-house
-Georgia;GA House District 48	state-house
-Georgia;GA House District 49	state-house
-Georgia;GA House District 50	state-house
-Georgia;GA House District 51	state-house
-Georgia;GA House District 52	state-house
-Georgia;GA House District 53	state-house
-Georgia;GA House District 54	state-house
-Georgia;GA House District 55	state-house
-Georgia;GA House District 56	state-house
-Georgia;GA House District 57	state-house
-Georgia;GA House District 58	state-house
-Georgia;GA House District 59	state-house
-Georgia;GA House District 60	state-house
-Georgia;GA House District 61	state-house
-Georgia;GA House District 62	state-house
-Georgia;GA House District 63	state-house
-Georgia;GA House District 64	state-house
-Georgia;GA House District 65	state-house
-Georgia;GA House District 66	state-house
-Georgia;GA House District 67	state-house
-Georgia;GA House District 68	state-house
-Georgia;GA House District 69	state-house
-Georgia;GA House District 70	state-house
-Georgia;GA House District 71	state-house
-Georgia;GA House District 72	state-house
-Georgia;GA House District 73	state-house
-Georgia;GA House District 74	state-house
-Georgia;GA House District 75	state-house
-Georgia;GA House District 76	state-house
-Georgia;GA House District 77	state-house
-Georgia;GA House District 78	state-house
-Georgia;GA House District 79	state-house
-Georgia;GA House District 80	state-house
-Georgia;GA House District 81	state-house
-Georgia;GA House District 82	state-house
-Georgia;GA House District 83	state-house
[…]
