+++ src/jurisdictions/Georgia/Candidate.txt
@@ -1,406 +1,1140 @@
 BallotName
-Zulma Lopez
-Zachary Perry
-Yasmin Neal
-Winfred Dukes
-William Park Freeman
-William K. Boddie, Jr.
-William Harris
-William D. Hughes
-William 'Bill' Werkheiser
-Will Wade
-Wes Cantrell
-Viola Davis
-Victor E. Anderson
-Veronica Brinson
-Vanessa Parker
-Vance Smith, Jr.
-Valencia Stovall
-Valencia M. Seay
-Val Almonord
-Tyler Paul Smith
-Tyler Harper
-Trey Rhodes
-Trey Kelley
-Treva Gear
-Travis Johnson
-Tracy Taylor
-Tonya P. Anderson
-Tommy Benton
-Tom Kirby
-Todd Jones
-Timothy Barr
-Tianna Smith
-Terry England
-Teri Anulewicz
-Taryn Chilivis Bowman
-Tamara Johnson-Shealey
-Taeho Cho
-Tabitha Johnson-Green
-Susan Holmes
-Steven Sainz
-Steven Meeks
-Steve Tarvin
-Steve Gooch
-Stephen M. George, Jr.
-Stephen Baughier
-Stan Gunter
-Stacey Evans
-Spencer Frye
-Soo Hong
-Sheri Smallwood Gilligan
-Shelly Hutchinson
-Sheila McNeill
-Sheila Jones
-Sheila Henley
-Sheila Clark Nelson
-Sheikh Rahman
-Shea Roberts
-Shaw Blackmon
-Sharonda Bell
-Sharon L. Ravert
-Sharon Henderson
-Sharon Cooper
-Shane Hazel
-Scott Holcomb
-Sarah Beeson
-Sara Tindall Ghazal
-Sanford Bishop
-Sandra Givens Scott
-Sam Watson
-Sam Park
-Sally Harrell
-Russ Goodman
-Rose Wing
-Ron Stephens
-Roger Bruce
-Robert Pruitt
-Robert Dickey
-Rob Leverett
-Ricky 'Rick' Williams
-Rick W. Allen
-Rick Jasperse
-Rick D. Day
-Richard H. Smith
-Richard Dien Winfield
-Rich McCormick
-Rhonda S. Taylor
-Rhonda Burnough
-Renitta Shannon
-Regina Lewis-Ward
-Regina Awung
-Rebecca Mitchell
-Raphael Warnock
-Randy Robertson
-Randy Nix
-Quentin T. Howell
-Q. Dailey
-Priscilla G. Smith
-Philip Singleton
-Pete Fuller
-Penny Houston
-Pedro 'Pete' Marin
-Patty Bentley
-Pat Rhudy
-Pat Calmes
-Park Cannon
-Pam Stephenson
-P. K. Martin, IV
-Noel Williams, Jr.
-Nikki Merritt
-Nikema Williams
-Nick Mitchell
-Natalie Bucsko
-Nan Orrock
-Nakita Hemingway
-Montenia Edwards
-Mitchell Scoggins
-Miriam Paris
-Mike Glanton
-Mike Dugan
-Mike Cheokas
-Mike Cameron
-Mickey Stephens
-Michelle Au
-Michael Todd Greene
-Michael Smith
-Michael S. Wilensky
-Michael Rhett
-Michael McConnell
-Michael Mack
-Michael 'Buckle' Moore
-Micah Gravley
-Mesha Mainor
-Melanie Williams
-Max Burns
-Matthew Wilson
-Matthew Gambill
-Matt Reeves
-Matt Lieberman
-Matt Hatchett
-Matt Dubnik
-Matt Dollar
-Matt Brass
[…]
