+++ src/jurisdictions/Georgia_BAK/Party.txt
@@ -1,6 +1,0 @@
-Name
-Democratic Party
-Republican Party
-Libertarian Party
-Green Party
-Independent Party

[…]
