### 2021-11-30 32922b8
Create sample session document (WIP)
