### 2021-12-02 259d7a6
documentation and some clean-up of ini, munger, juris files
