+++ src/jurisdictions/Minnesota/Party.txt
@@ -4,28 +4,16 @@ Libertarian Party
 American Delta Party
 Constitution Party
 Democratic-Farmer-Labor Party
-Grassroots - Legalize Cannabis
-Humanity Forward
-Independence
-Independent
-Independent - Judicial Reform
-Independent - unaffiliated
-Independent Progressive Liberal
-Legal Marijuana Now
-Green Party
-Nonpartisan
-Socialist Workers Party
-Unaffiliated
-Write-In
-VPM
-SLP
-IND
-IA
-I
 Grassroots-Legalize Cannabis Party
+Humanity Forward
+IA Party
+Independent - Judicial Reform Party
+Independent Progressive Liberal Party
+Legal Marijuana Now Party
+Green Party
+Socialist Workers Party
+VPM Party
+SLP Party
 Independence Party
 Independent Party
-Legal Marijuana Now Party
 Minnesota Green Party
-Nonpartisan Party
-none or unknown
