+++ requirements.txt
@@ -1,5 +1,5 @@
 matplotlib==3.1.2
-numpy==1.18.1
+numpy==1.21.4
 pandas==1.3.2
 psycopg2==2.8.6
 scipy==1.6.3
