+++ src/ini_files_for_results/Colorado/co20g_unofficial.ini
@@ -7,4 +7,4 @@ results_short_name=co20g
 results_download_date=2020-11-04
 results_source=CO SoS
 results_note=
-aux_data_dir=
+is_preliminary=True
