+++ src/ini_files_for_results/Missouri/MissouriUSCongressDistrict4_20g.ini
@@ -7,6 +7,6 @@ results_short_name=mo_20g_MissouriUSCongressDistrict4
 results_download_date=2020-11-05
 results_source=https://enrarchives.sos.mo.gov/enrnet/default.aspx?eid=750003949
 results_note=
-aux_data_dir=
+is_preliminary=True
 CountItemType=total
 CandidateContest=US House MO District 4
