+++ src/ini_files_for_results/Oregon/Hood River_OR_20g.ini
@@ -7,6 +7,6 @@ results_short_name=OR20gxlsx_Hood River
 results_download_date=2020-11-10
 results_source=https://results.oregonvotes.gov/ResultsExport.aspx
 results_note=
-aux_data_dir=
+is_preliminary=True
 ReportingUnit=Oregon;Hood River County
 CountItemType=total
