+++ tests/analyzer_tests/test_analyzer.py
@@ -70,7 +70,7 @@ def test_contest_type_display(analyzer):
         analyzer.display_options("contest_type", filters=["2018 General", "Georgia"]),
         sort_key="order_by",
     )
-    expected = dict_sort(results.ga_2018_contest_types, sort_key="order_by")
+    expected = dict_sort(results.ga_2018_contest_types,sort_key="order_by")
     assert new == expected
 
 
@@ -80,16 +80,16 @@ def test_contest_display(analyzer):
         filters=["2018 General", "Georgia", "Congressional"],
     )
     new = dict_sort(contests, sort_key="order_by")
-    expected = dict_sort(results.ga_2018_congressional_contests, sort_key="order_by")
+    expected = dict_sort(results.ga_2018_congressional_contests,sort_key="order_by")
     assert new == expected
 
 
 def test_bar_congressional(analyzer):
     assert (
-        analyzer.bar(
+            analyzer.bar(
             "2018 General", "Georgia", "Congressional", "US House GA District 3"
         )
-        == results.ga_2018_bar_congressional
+            == results.ga_2018_bar_congressional
     )
 
 
@@ -111,17 +111,17 @@ def test_bar_congressional(analyzer):
 
 def test_bar_all_state(analyzer):
     assert (
-        analyzer.bar("2018 General", "North Carolina", "State House", "All State House")
-        == results.nc_2018_bar_statehouse
+            analyzer.bar("2018 General", "North Carolina", "State House", "All State House")
+            == results.nc_2018_bar_statehouse
     )
 
 
 def test_bar_all_congressional(analyzer):
     assert (
-        analyzer.bar(
+            analyzer.bar(
             "2018 General", "North Carolina", "Congressional", "All Congressional"
         )
-        == results.nc_2018_bar_congressional
+            == results.nc_2018_bar_congressional
     )
 
 
@@ -133,7 +133,7 @@ def test_category_display(analyzer):
         ),
         sort_key="order_by",
     )
-    expected = dict_sort(results.nc_2018_category, sort_key="order_by")
+    expected = dict_sort(results.nc_2018_category,sort_key="order_by")
     assert new == expected
 
 
@@ -145,13 +145,13 @@ def test_count_display(analyzer):
         ),
         sort_key="order_by",
     )
-    expected = dict_sort(results.ga_2018_count, sort_key="order_by")
+    expected = dict_sort(results.ga_2018_count,sort_key="order_by")
     assert new == expected
 
 
 def test_scatter_candidates(analyzer):
     assert (
-        analyzer.scatter(
+            analyzer.scatter(
             "Georgia",
             "2018 General",
             "Candidate total",
@@ -160,13 +160,13 @@ def test_scatter_candidates(analyzer):
             "Candidate total",
             "Charlie Bailey",
         )
-        == results.ga_2018_scatter_candidates
+            == results.ga_2018_scatter_candidates
     )
 
 
 def test_scatter_candidates_longname(analyzer):
     assert (
-        analyzer.scatter(
+            analyzer.scatter(
             "Georgia",
             "2018 General",
             "Candidate total",
@@ -175,13 +175,13 @@ def test_scatter_candidates_longname(analyzer):
             "Candidate total",
             "Charlie Bailey - D - GAAttGen",
         )
-        == results.ga_2018_scatter_candidates
+            == results.ga_2018_scatter_candidates
     )
 
 
 def test_scatter_candidates_votetype(analyzer):
     assert (
-        analyzer.scatter(
+            analyzer.scatter(
             "Georgia",
             "2018 General",
             "Candidate absentee-mail",
@@ -190,13 +190,13 @@ def test_scatter_candidates_votetype(analyzer):
             "Candidate election-day",
             "Stacey Abrams",
[…]
