+++ docs/Testing_Code_with_pytest.md
@@ -7,7 +7,7 @@ The dataloading tests rely on having some raw results data to load. And the resu
  Call the tests from a working directory with the following structure and files:
 ```
 .
-+-- input_directory
++-- input_results
 |   +-- Alabama
 |   |   + <Alabama results file>
 |   |   + <maybe another Alabama results file>
@@ -17,22 +17,18 @@ The dataloading tests rely on having some raw results data to load. And the resu
 |   |   + <American Samoa results file>
 |   +-- <etc>
 |
-+-- reports_and_plots_directory
++-- reports_and_plots
 +-- run_time.ini
 ```
-where the contents of run_time.ini are:
-```
-[electiondata]
-results_dir=input_directory
-archive_dir=archive_directory
-repository_content_root=<path/to/src>
-reports_and_plots_dir=reports_and_plots_directory
+The file `run_time.ini` can be the same as in the [Sample Dataloading Session](Sample_Dataloading_Session.md). 
 
-[postgresql]
-host=localhost
-port=5432
-dbname=pytest
-user=postgres
-password=
-```
- 
+## Note on dataloading tests
+The tests in [test_dataloading_by_ej.py](../tests/dataloading_tests/test_dataloading_by_ej.py) will attempt to load all raw results files in `input_results` that are specified by some file in the [`ini_file_for_results` directory](../src/ini_files_for_results). You can check which jurisdictions had files loaded:
+ * if the test is successful, look at the `compare_*` directories in the `reports_and_plots` directory.
+ * if the test fails, look in the output from the test.
+
+## Running the tests
+You will need pytest to be installed on your system (see [pytest installation instructions](https://docs.pytest.org/en/6.2.x/getting-started.html) if necessary). Commands are run from the shell
+ * dataloading routines: `pytest ~/PycharmProjects/electiondata/tests/dataloading_tests`
+ * jurisdiction prep routines: `pytest ~/PycharmProjects/electiondata/tests/jurisdiction_prepper_tests/`
+ * analysis routines: `pytest ~/PycharmProjects/electiondata/tests/analyzer_tests/  `
