### 2021-12-07 6bf1379
add encoding=constants.default_encoding in report()
