### 2021-12-07 37d54ee
add encoding=constants.default_encoding in report()
