### 2021-12-07 #3739 @sfsinger19103
@vaneseltine How might I test my fix to the encoding issue, which is in branch `issue739-windows-encoding`?
