+++ src/electiondata/userinterface/__init__.py
@@ -8,21 +8,22 @@ from configparser import (
 from electiondata import (
     database as db,
     munge as m,
-    juris as jm,
     nist as nist,
     constants,
 )
+
+from slugify import slugify
 import pandas as pd
 from pandas.errors import ParserError
-import os
+from os import walk, listdir
+from os.path import join, isdir, isfile
 from pathlib import Path
 from typing import Optional, Dict, Any, List
-import datetime
-import csv
-import numpy as np
-import inspect
-import xml.etree.ElementTree as et
-import json
+from csv import QUOTE_MINIMAL
+from numpy import where 
+from inspect import currentframe
+from xml.etree.ElementTree import parse
+from json import loads
 import shutil
 import xlrd
 
@@ -191,7 +192,7 @@ def read_single_datafile(
         kwargs = tabular_kwargs(p, kwargs, aux=aux)
         if p["multi_block"] == "yes":
             kwargs["header"] = None
-        kwargs["quoting"] = csv.QUOTE_MINIMAL
+        kwargs["quoting"] = QUOTE_MINIMAL
         if p["flat_text_delimiter"] in ["tab", "\\t"]:
             kwargs["sep"] = "\t"
         else:
@@ -207,7 +208,7 @@ def read_single_datafile(
             else:
                 driver = nist.xml_count_parse_info(p, ignore_namespace=True)
             xml_path_info = nist.xml_string_path_info(p["munge_fields"], p["namespace"])
-            tree = et.parse(f_path)
+            tree = parse(f_path)
             df, err = nist.df_from_tree(
                 tree,
                 xml_path_info=xml_path_info,
@@ -221,7 +222,7 @@ def read_single_datafile(
         elif p["file_type"] in ["json-nested"]:
             # TODO what if json-nested is a lookup?
             with open(f_path, "r") as f:
-                data = json.loads(f.read())
+                data = loads(f.read())
             df = pd.json_normalize(data, **kwargs)
             if not fatal_error(err):
                 df.rename(columns=rename, inplace=True)
@@ -273,7 +274,7 @@ def read_single_datafile(
                     err = add_new_error(
                         err,
                         "system",
-                        f"{Path(__file__).absolute().parents[0].name}.{inspect.currentframe().f_code.co_name}",
+                        f"{Path(__file__).absolute().parents[0].name}.{currentframe().f_code.co_name}",
                         f"Unexpected error setting and filling headers after padding file {file_name}",
                     )
 
@@ -366,7 +367,7 @@ def excel_to_dict(
         err = add_new_error(
             err,
             "system",
-            f"{Path(__file__).absolute().parents[0].name}.{inspect.currentframe().f_code.co_name}",
+            f"{Path(__file__).absolute().parents[0].name}.{currentframe().f_code.co_name}",
             f"Unexpected exception while getting row-constant keyword arguments for \n"
             f"rows_to_read: {rows_to_read}\n"
             f"kwargs: {kwargs}.\n"
@@ -455,15 +456,15 @@ def copy_directory_with_backup(
     <backup_suffix>"""
     err = None
     # if the original to be copied is actually a directory
-    if os.path.isdir(original_path):
+    if isdir(original_path):
         if backup_suffix:
             # make backup of anything with existing name
-            if os.path.isdir(copy_path):
+            if isdir(copy_path):
                 shutil.move(copy_path, f"{copy_path}{backup_suffix}")
                 print(f"Moved {copy_path} to {copy_path}{backup_suffix}")
-            elif os.path.isfile(copy_path):
+            elif isfile(copy_path):
                 old_stem = Path(copy_path).stem
-                backup_path = os.path.join(
+                backup_path = join(
                     Path(copy_path).parent,
                     f"{old_stem}{backup_suffix}.{Path(copy_path).suffix}",
                 )
[…]
