### 2021-12-08 09be2e7
slugify output file names; make import statements more efficient
