+++ src/electiondata/userinterface/__init__.py
@@ -13,7 +13,7 @@ from electiondata import (
 )
 
 from slugify import slugify
-import pandas as pd
+from pandas import DataFrame, ExcelFile, json_normalize, read_csv, MultiIndex, read_excel, concat
 from pandas.errors import ParserError
 from os import walk, listdir
 from os.path import join, isdir, isfile
@@ -24,15 +24,9 @@ from numpy import where
 from inspect import currentframe
 from xml.etree.ElementTree import parse
 from json import loads
-import shutil
-import xlrd
-
-# may need for certain excel imports: import openpyxl
+from shutil import move, copy
+from xlrd import open_workbook
 from sqlalchemy.orm import Session
-
-
-# mapping from internal database reportingunit types to the user-facing contest types
-# (contests are categorized by the reporting unit type of their corresponding districts)
 
 
 def find_dupes(df):
@@ -60,9 +54,9 @@ def json_kwargs(
         else:
             json_rename[path_list[-1]] = mf
     meta = list(list(t) for t in meta_set)
-    json_kwargs = {"meta": meta, "record_path": record_path, "errors": "ignore"}
+    j_kwargs = {"meta": meta, "record_path": record_path, "errors": "ignore"}
 
-    return json_kwargs, json_rename
+    return j_kwargs, json_rename
 
 
 def tabular_kwargs(
@@ -134,14 +128,14 @@ def list_desired_excel_sheets(f_path: str, p: dict) -> (Optional[list], Optional
         try:
             # read an xlsx file
             # # nb: the following fails on VT 2020 files
-            xl = pd.ExcelFile(f_path)
+            xl = ExcelFile(f_path)
             all_sheets = xl.sheet_names
             # xlsx = openpyxl.load_workbook(f_path)
             # all_sheets = xlsx.get_sheet_names()
-        except Exception as exc:
+        except Exception:
             try:
                 # read xls file
-                xls = xlrd.open_workbook(f_path, on_demand=True)
+                xls = open_workbook(f_path, on_demand=True)
                 all_sheets = xls.sheet_names()
             except Exception as exc:
                 err = add_new_error(
@@ -170,7 +164,7 @@ def read_single_datafile(
     aux: bool = False,
     driving_path: Optional[str] = None,
     lookup_id: Optional[str] = None,
-) -> (Dict[str, pd.DataFrame], Dict[str, Dict[int, Any]], Optional[dict]):
+) -> (Dict[str, DataFrame], Dict[str, Dict[int, Any]], Optional[dict]):
     """Length of returned dictionary is the number of sheets read -- usually 1 except for multi-sheet Excel.
     Auxiliary files have different parameters (e.g., no count locations)"""
     err = None
@@ -223,7 +217,7 @@ def read_single_datafile(
             # TODO what if json-nested is a lookup?
             with open(f_path, "r") as f:
                 data = loads(f.read())
-            df = pd.json_normalize(data, **kwargs)
+            df = json_normalize(data, **kwargs)
             if not fatal_error(err):
                 df.rename(columns=rename, inplace=True)
                 df_dict = {"Sheet1": df}
@@ -245,7 +239,7 @@ def read_single_datafile(
                 df_dict = dict()
         elif p["file_type"] == "flat_text":
             try:
-                df = pd.read_csv(f_path, **kwargs)
+                df = read_csv(f_path, **kwargs)
             except ValueError as ve:
                 print(
                     f"ValueError (while reading flat text file), possibly from uneven record lengths: {ve}\n "
@@ -256,7 +250,7 @@ def read_single_datafile(
                 kwargs_pad = kwargs
                 kwargs_pad["index_col"] = None
                 kwargs_pad["header"] = None
[…]
