+++ requirements.txt
@@ -13,3 +13,4 @@ requests==2.25.1
 openpyxl==3.0.6
 dicttoxml==1.7.4
 lxml==4.6.3
+slugify=5.0.3
