### 2021-12-08 #3739 @vaneseltine
> @vaneseltine How might I test my fix to the encoding issue, which is in branch `issue739-windows-encoding`?

Hmm. What you'd like to do is have an instance of Python with a "wrong" default codec (such as 1252) attempt to run using a file encoded as UTF-8. I've been looking around for a little while this afternoon and, due to Python's increasing reluctance to allow changing of locale "defaults," I'm not sure there's a good way to create a straightforward, independent unit test.

One reasonably simple approach that I _think_ should work is to open a temporary environment with the environment variable PYTHONIOENCODING = 1252. That might be still be possible inside Python somehow, but a shell script would be the more clear route.

Then pytest to load_all() on the TestingData repo and check for UnicodeEncodeError, though for efficiency's sake, you'll probably want to narrow down to the subset that causes the problem. (And, of course, you'll want to make sure that your setup is properly crashing when using the main branch.)
