+++ src/electiondata/userinterface/__init__.py
@@ -4,29 +4,28 @@ from configparser import (
     DuplicateOptionError,
     ParsingError,
 )
+from csv import QUOTE_MINIMAL
+from inspect import currentframe
+import json
+from numpy import where
+from os import walk, listdir
+from os.path import join, isdir, isfile
+import pandas as pd
+from pathlib import Path
+import shutil
+from slugify import slugify
+from sqlalchemy.orm import Session
+from typing import Optional, Dict, Any, List
+import xlrd
+import xml.etree.ElementTree
 
+# local imports
 from electiondata import (
     database as db,
     munge as m,
     nist as nist,
     constants,
 )
-
-from slugify import slugify
-from pandas import DataFrame, ExcelFile, json_normalize, read_csv, MultiIndex, read_excel, concat
-from pandas.errors import ParserError
-from os import walk, listdir
-from os.path import join, isdir, isfile
-from pathlib import Path
-from typing import Optional, Dict, Any, List
-from csv import QUOTE_MINIMAL
-from numpy import where 
-from inspect import currentframe
-from xml.etree.ElementTree import parse
-from json import loads
-from shutil import move, copy
-from xlrd import open_workbook
-from sqlalchemy.orm import Session
 
 
 def find_dupes(df):
@@ -128,14 +127,14 @@ def list_desired_excel_sheets(f_path: str, p: dict) -> (Optional[list], Optional
         try:
             # read an xlsx file
             # # nb: the following fails on VT 2020 files
-            xl = ExcelFile(f_path)
+            xl = pd.ExcelFile(f_path)
             all_sheets = xl.sheet_names
             # xlsx = openpyxl.load_workbook(f_path)
             # all_sheets = xlsx.get_sheet_names()
         except Exception:
             try:
                 # read xls file
-                xls = open_workbook(f_path, on_demand=True)
+                xls = xlrd.open_workbook(f_path, on_demand=True)
                 all_sheets = xls.sheet_names()
             except Exception as exc:
                 err = add_new_error(
@@ -164,7 +163,7 @@ def read_single_datafile(
     aux: bool = False,
     driving_path: Optional[str] = None,
     lookup_id: Optional[str] = None,
-) -> (Dict[str, DataFrame], Dict[str, Dict[int, Any]], Optional[dict]):
+) -> (Dict[str, pd.DataFrame], Dict[str, Dict[int, Any]], Optional[dict]):
     """Length of returned dictionary is the number of sheets read -- usually 1 except for multi-sheet Excel.
     Auxiliary files have different parameters (e.g., no count locations)"""
     err = None
@@ -202,7 +201,7 @@ def read_single_datafile(
             else:
                 driver = nist.xml_count_parse_info(p, ignore_namespace=True)
             xml_path_info = nist.xml_string_path_info(p["munge_fields"], p["namespace"])
-            tree = parse(f_path)
+            tree = xml.etree.ElementTree.parse(f_path)
             df, err = nist.df_from_tree(
                 tree,
                 xml_path_info=xml_path_info,
@@ -216,8 +215,8 @@ def read_single_datafile(
         elif p["file_type"] in ["json-nested"]:
             # TODO what if json-nested is a lookup?
             with open(f_path, "r") as f:
-                data = loads(f.read())
-            df = json_normalize(data, **kwargs)
+                data = json.loads(f.read())
+            df = pd.json_normalize(data, **kwargs)
             if not fatal_error(err):
                 df.rename(columns=rename, inplace=True)
                 df_dict = {"Sheet1": df}
@@ -239,7 +238,7 @@ def read_single_datafile(
                 df_dict = dict()
         elif p["file_type"] == "flat_text":
             try:
-                df = read_csv(f_path, **kwargs)
+                df = pd.read_csv(f_path, **kwargs)
[…]
