### 2021-12-09 fe5ebc1
 This is a combination of 2 commits.  This is the 1st commit message:
Add community guidelines

 The commit message #2 will be skipped:

 Add conduct file
