+++ CONTRIBUTING.md
@@ -14,81 +14,47 @@ you agree to abide by its terms.
 
 When filing an issue, the most important thing is to include a minimal 
 reproducible example so that we can quickly verify the problem, and then figure 
-out how to fix it. There are three things you need to include to make your 
-example reproducible: required packages, data, code.
+out how to fix it. Please include
 
-1.  **Packages** should be loaded at the top of the script, so it's easy to
-    see which ones the example needs.
+1.  Link to the exact version of electiondata you are using.
   
-1.  The easiest way to include **data** is to use `dput()` to generate the R code 
-    to recreate it. For example, to recreate the `mtcars` dataset in R,
-    I'd perform the following steps:
+1.  A copy of your working directory, following the model in the [Sample Session](docs/Sample_Session.md). Include:
+    * a `run_time.ini` file (Feel free to redact the login information for your local postgres instance.) 
+    * all subdirectories referenced in the `run_time.ini` file. 
+    * If the issue involves loading a particular results data file, be sure to include that file. If possible, avoid submitting large results files -- if your file is precinct-based, for example, see if you can demonstrate the issue with a truncated results file.
+    * all error and warnings files placed by the system into the `reports_and_plots_dir` specified in `run_time.ini`.
   
-       1. Run `dput(mtcars)` in R
-       2. Copy the output
-       3. In my reproducible script, type `mtcars <- ` then paste.
-       
-    But even better is if you can create a `data.frame()` with just a handful
-    of rows and columns that still illustrates the problem.
-  
-1.  Spend a little bit of time ensuring that your **code** is easy for others to
-    read:
-  
-    * make sure you've used spaces and your variable names are concise, but
-      informative
-  
-    * use comments to indicate where your problem lies
-  
-    * do your best to remove everything that is not related to the problem.  
-     The shorter your code is, the easier it is to understand.
+1.  A transcript of the python session, where python is called from the working directory.
 
-You can check you have actually made a reproducible example by starting up a 
-fresh R session and pasting your script in.
-
-(Unless you've been specifically asked for it, please don't include the output 
-of `sessionInfo()`.)
+You can check you have actually made a reproducible example by:
+1. creating a virtual environment
+1. installing the indicated version of electiondata
+1. if files or folders were moved by the system to the archive directory, move them back to the input directory, removing any timestamps from directory names.
+1. navigating to the working directory, calling python, and producing the behavior in question.
 
 ## Pull requests
 
-To contribute a change to ggplot2, you follow these steps:
+To contribute a change to `electiondata` follow these steps:
 
 1. Create a branch in git and make your changes.
+1. Run unit tests to check for broken functionality. See [pytest instructions](docs/Testing_Code_with_pytest.md).
+1. Run python `black` to format your code to our standard.
 1. Push branch to github and issue pull request (PR).
 1. Discuss the pull request.
 1. Iterate until either we accept the PR or decide that it's not
-   a good fit for ggplot2.
+   a good fit for `electiondata`.
 
 Each of these steps are described in more detail below. This might feel 
 overwhelming the first time you get set up, but it gets easier with practice. 
-If you get stuck at any point, please reach out for help on the [ggplot2-dev](https://groups.google.com/forum/#!forum/ggplot2-dev) mailing list.
+If you get stuck at any point, feel free to [contact us](CONTACT_US.md) for help.
 
-If you're not familiar with git or github, please start by reading <http://r-pkgs.had.co.nz/git.html>
+If you're not familiar with git or github, please read a tutorial such as [https://realpython.com/python-git-github-intro/](https://realpython.com/python-git-github-intro/).
 
-<!--
[…]
