### 2021-12-13 #3739 @sfsinger19103
@vaneseltine I tried a virtual environment with PYTHONIOENCODING = 1252 in the `main` branch, but couldn't get the data loading to fail the way it did on your windows system. I see from your output that the failure happened in an Alaska file. It would be useful to know which one to ensure that I tested on that one specifically.

Any chance you could try again on your system using the version in the `Issue739-windows-encoding` branch? If that doesn't work I'm not sure what to do next.
