### 2021-12-13 #3739 @sfsinger19103
And note @vaneseltine that I've removed some of the data from the TestingData repo, to get rid of jurisdictions (like Philadelphia) and some files (from elections before the 2020 General) that don't load correctly in the current version.
