### 2021-12-13 b4a1861
add bar charts to Sample Session, fixing some bugs found along the way; Change name from Sample_Dataloading_Session.md to Sample_Session.md
