### 2021-12-13 ecf59f2
improve file names for exported images
