### 2021-12-13 4158508
delete obsolete/incorrect NIST cdf import info
