+++ docs/User_Guide.md
@@ -1,5 +1,4 @@
-# How to Use the System
-## Environment
+# Environment
 You will need:
  * `python3.9`. 
  * `postgresql`. You do *not* need to create a database; you just need to make sure a `postgresql` is installed, and you need to have the hostname, port and login credentials with default permissions. 
@@ -9,20 +8,26 @@ For more detail, see the [Installation Instructions](./Installation.md).
 
 To use other varieties of SQL, you would need to modify the routines in the `database` module.
 
-## Installation
+# Installation
 The one-line version: From the root folder of your repository run `python3 setup.py install`. For more detail, see the [Installation Instructions](./Installation.md).
 
-### Main parameter file
+## Main parameter file
 You will need a main parameter file to specify paths and database connection information specific to your local computing environment. This file is necessary for the three main classes:
  * `JurisdictionPrepper` for preparing jurisdiction files
  * `DataLoader` for loading data
  *  `Analyzer` for exporting and analyzing results
 See the [template file](../src/parameter_file_templates/run_time.ini.template) for required parameters. Avoid percent signs and line breaks in the parameter values.
    
-### Other recommended files
+## Other recommended files
 To avoid the overhead of deriving the major subdivision type for each jurisdiction from the database, make sure that your repository has a [000_major_subjurisdiction_types.txt](../src/jurisdictions/000_for_all_jurisdictions/major_subjurisdiction_types.txt) in the [jurisdictions directory](../src/jurisdictions/). This file allows the user to specify other major subdivisions. For example, it may make sense to consider towns as the major subdivisions in Connecticut rather than counties. Or a user may wish to use congressional districts as the major subdivision -- though such a user should not assume that the nesting relationships (say, of precincts within congressional districts) have been coded in the [`ReportingUnit.txt` file](../src/jurisdictions/Connecticut/ReportingUnit.txt) or the database.
 
-## Determining a Munger
+# Data Load Preparation
+In order to process a raw results data file, the system needs specified:
+   * a "munger" with file format details -- e.g., where to find the candidate name associated with a particular vote count 
+   * a "dictionary.txt" file to translate from naming conventions of the file into the naming conventions for the database
+   * other "jurisdiction files" listing legitimate contests, candidates, etc., with information about each (e.g. election district or party)
+
+## Mungers
 Election result data comes in a variety of file formats. Even when the basic format is the same, file columns may have different interpretations. The code is built to ease -- as much as possible -- the chore of processing and interpreting each format. Following the [Jargon File](http://catb.org/jargon/html/M/munge.html), which gives one meaning of "munge" as "modify data in some way the speaker doesn't need to go into right now or cannot describe succinctly," we call each set of basic information about interpreting an election result file a "munger". 
 
 If the munger for the format of your results file doesn't already exist:
@@ -89,7 +94,7 @@ WARD,DIVISION,VOTE TYPE,CATEGORY,SELECTION,PARTY,VOTE COUNT
 ```
 The formula `ReportingUnit=Ward <WARD>;Division <DIVISION>` would yield 'Ward 01;Division 01'.
 
-#### the `count_location` parameter for xml and json files
+### the `count_location` parameter for xml and json files
 In tree-structured file types like json and xml files, the `count_location` parameter is a path to the count.
 
 Consider this snippet from a Georgia voting results xml file:
@@ -135,7 +140,7 @@ Each pair of braces ({}) and its contents is a json "object", while each pair of
 Note that `count_location` is a `/`-separated path, possibly with a period (`.`) at the end for xml files where the count is in an attribute. Munge formulas use only periods.
 
 
-#### flat text 
+### flat text 
[…]
