### 2021-12-14 04d2d2e
add reference for bar chart selection algorithm
