### 2021-12-16 #3739 @sfsinger19103
Thank you for all this great feedback. I believe I have addressed it all. I believe I have fixed the windows encoding issue, but was unable to test it definitively (see above).

@ajstewartlang Please let me know if you need anything else from me.
