### 2021-12-16 #3739 @sfsinger19103
@vaneseltine just realized I hadn't merged the encoding issue fix into `feature-version-2.0.1`. Will do that in the next hour or so.
