### 2021-12-16 a7e2f47
document; remove top_counts method
