+++ src/electiondata/analyze/__init__.py
@@ -1,151 +1,79 @@
-import os.path
-from typing import Optional, List
-from sqlalchemy.orm import Session
-import psycopg2
-import pandas as pd
 import inspect
+import json
+import numpy as np
+import pandas as pd
+from pathlib import Path
+import psycopg2
+from scipy import stats
+import scipy.spatial.distance as dist
+from sqlalchemy.orm import Session
+from typing import Any, Dict, Optional, List
+
 from electiondata import (
     userinterface as ui,
     munge as m,
     database as db,
     constants,
 )
-import datetime
-import os
-import numpy as np
-from pathlib import Path
-import scipy.spatial.distance as dist
-from scipy import stats
-import json
-
-
-def child_rus_by_id(session, parents, ru_type: str = None):
-    """Given a list <parents> of parent ids (or just a single parent_id), return
-    list containing all children of those parents.
-    (By convention, a ReportingUnit counts as one of its own 'parents',)
-    If  <rutype> is given,
-    restrict children to that ReportingUnitType"""
-    cruj = pd.read_sql_table("ComposingReportingUnitJoin", session.bind)
-    children = list(
-        cruj[cruj.ParentReportingUnit_Id.isin(parents)].ChildReportingUnit_Id.unique()
-    )
-    if ru_type:
-        assert len(ru_type) == 2, f"argument {ru_type} does not have exactly 2 elements"
-        ru = pd.read_sql_table("ReportingUnit", session.bind, index_col="Id")
-        right_type_ru = ru[(ru.ReportingUnitType == ru_type)]
-        children = [x for x in children if x in right_type_ru.index]
-    return children
-
-
-def export_rollup(
-    session,
-    target_dir: str,
-    jurisdiction_id: int,
-    sub_rutype: str,
-    election_id: int,
-    datafile_list: list = None,
-    by: str = "Id",
-    by_vote_type: bool = False,
-) -> str:
-    """<target_dir> is the directory where the resulting rollup_dataframe will be stored.
-    <election_id> identifies the election; <datafile_id_list> the datafile whose results will be rolled up.
-    <top_ru_id> is the internal cdf name of the ReportingUnit whose results will be reported
-    <sub_rutype> identifies the ReportingUnitType
-    of the ReportingUnits used in each line of the results file
-    created by the routine. (E.g., county or ward)
-    <datafile_list> is a list of files, with entries from field <by> in _datafile table.
-    If no <datafile_list> is given, return all results for the given election.
-    """
-
-    connection = session.bind.raw_connection()
-    cursor = connection.cursor()
-    if not datafile_list:
-        datafile_list, e = db.data_file_list_cursor(cursor, election_id, by="Id")
-        if e:
-            return e
-        by = "Id"
-        if len(datafile_list) == 0:
-            return f"No datafiles found for Election_Id {election_id}"
-    # set exclude_redundant_total
-    vote_type_list, err_str = db.vote_type_list(cursor, datafile_list, by=by)
-    if err_str:
-        return err_str
-    elif len(vote_type_list) == 0:
-        return f"No vote types found for datafiles with {by} in {datafile_list} "
-
-    if len(vote_type_list) > 1 and "total" in vote_type_list:
-        exclude_redundant_total = True
-    else:
-        exclude_redundant_total = False
-
-    # get names from ids (and sub_rutype othertext if appropriate)
-    top_ru = db.name_from_id_cursor(cursor, "ReportingUnit", jurisdiction_id)
-    election = db.name_from_id_cursor(cursor, "Election", election_id)
-
-    # create path to export directory
[…]
