### 2021-12-18 #3739 @vaneseltine
Great! I just finished running my way through the complete TestingData repo without any encoding or filename issues.

It doesn't look like there will be an straightforward automated way to test for these problems without a Windows machine. In the future, if others report Windows issues, you could consider an automated CI/runner/task platform with Windows images available. GitHub Actions is a pretty easy system that you could configure to spin up multiple OS environments to run your pytests whenever you commit.

@ajstewartlang -- my review is complete and the points I raised have been addressed. This is good to go by me.
