+++ src/electiondata/analyze/__init__.py
@@ -3,7 +3,6 @@ import json
 import numpy as np
 import pandas as pd
 from pathlib import Path
-import psycopg2
 from scipy import stats
 import scipy.spatial.distance as dist
 from sqlalchemy.orm import Session
@@ -74,8 +73,6 @@ def create_scatter(
             "x": string, shorthand identifier for horizontal data
             "y": string, shorthand identifier for vertical axis
     """
-    connection = session.bind.raw_connection()
-    cursor = connection.cursor()
 
     # get the mappings back to the DB labels
     h_count = ui.get_contest_type_mapping(h_count)
@@ -102,11 +99,10 @@ def create_scatter(
         v_runoff,
     )
     if dfh.empty or dfv.empty:
-        connection.close()
         return None
 
     unsummed = pd.concat([dfh, dfv])
-    jurisdiction = db.name_from_id_cursor(cursor, "ReportingUnit", jurisdiction_id)
+    jurisdiction = db.name_from_id(session, "ReportingUnit", jurisdiction_id)
 
     # check if there is only 1 candidate selection (with multiple count types)
     single_selection = len(unsummed["Selection"].unique()) == 1
@@ -127,7 +123,6 @@ def create_scatter(
     pivot_df = pivot_df.dropna()
     pivot_df.columns = pivot_df.columns.map(str)
     if pivot_df.empty:
-        connection.close()
         return None
 
     # package up results
@@ -148,27 +143,27 @@ def create_scatter(
         results["y"] = v_count
     else:  # neither is runoff; not single_selection
         results = package_results(pivot_df, jurisdiction, h_count, v_count)
-    results["x-election"] = db.name_from_id_cursor(cursor, "Election", h_election_id)
-    results["y-election"] = db.name_from_id_cursor(cursor, "Election", v_election_id)
+    results["x-election"] = db.name_from_id(session, "Election", h_election_id)
+    results["y-election"] = db.name_from_id(session, "Election", v_election_id)
     results["subdivision_type"] = subdivision_type
     results["x-count_item_type"] = h_category
     results["y-count_item_type"] = v_category
     results["x-title"] = scatter_axis_title(
-        cursor,
+        session,
         results["x"],
         results["x-election"],
         dfh.iloc[0]["Contest"],
         jurisdiction_id,
     )
     results["y-title"] = scatter_axis_title(
-        cursor,
+        session,
         results["y"],
         results["y-election"],
         dfv.iloc[0]["Contest"],
         jurisdiction_id,
     )
-    h_preliminary = db.is_preliminary(cursor, h_election_id, jurisdiction_id)
-    v_preliminary = db.is_preliminary(cursor, v_election_id, jurisdiction_id)
+    h_preliminary = db.is_preliminary(session,h_election_id,jurisdiction_id)
+    v_preliminary = db.is_preliminary(session,v_election_id,jurisdiction_id)
     results["preliminary"] = h_preliminary or v_preliminary
 
     # only keep the ones where there are an (x, y) to graph
@@ -179,11 +174,9 @@ def create_scatter(
         if len(result) == 5:
             to_keep.append(result)
     if not to_keep:
-        connection.close()
         return None
 
     results["counts"] = to_keep
-    connection.close()
     return results
 
 
@@ -291,10 +284,8 @@ def get_external_data(
     subdivision_type,
 ):
     # get the census data
-    connection = session.bind.raw_connection()
-    cursor = connection.cursor()
     census_df = db.read_external(
-        cursor,
+        session,
         election_id,
         jurisdiction_id,
         ["Name", "Category", "Label", "Value", "Source"],
@@ -302,7 +293,6 @@ def get_external_data(
         restrict_by_category=category,
         subdivision_type=subdivision_type,
[…]
