### 2021-12-19 3b9fb2d
remove psycopg2 references outside database module
