### 2021-12-19 e5783ce
delete obsolete database.vote_type_list function
