### 2021-12-19 e5adbc8
document inline (WIP)
