+++ src/electiondata/analyze/__init__.py
@@ -238,20 +238,25 @@ def get_data_for_scatter(
     subdivision_type: str,
     election_id: int,
     count_item_type: str,
-    filter_str,
-    count_type,
-    is_runoff,
-):
+    filter_str: str,
+    count_type: str,
+    is_runoff: bool,
+) -> pd.DataFrame:
     """
     :param session: sqlalchemy database session
     :param jurisdiction_id: integer Id of jurisdiction in database ReportingUnit table
-    :param subdivision_type: string ReportingUnitType characterizing subdivisions for points in scatter
+    :param subdivision_type: string ReportingUnitType
     :param election_id: integer Id of election in database Election table
     :param count_item_type: CountItemType characterizing vote counts
-    :param filter_str:
-    :param count_type:
-    :param is_runoff:
-    :return:
+    :param filter_str: string used to filter data
+        for vote counts: "All contests" or "All candidates" or a value for Contest.Name or Candidate.BallotName
+        for other datasets: value for ExternalDataSet.Label
+    :param count_type: string used to filter data:
+        for vote counts: "candidates" or "parties" or "contests"
+        for other datasets: value for ExternalDataSet.Category
+    :param is_runoff: True for a run-off contest associated to the election, otherwise False
+    :return: dataframe of results rolled up to ReportingUnits of <subdivision> ReportingUnitType,
+        restricted by <count_item_type>, <filter_str" and <count_type>
     """
     if count_type.startswith("Population"):
         return get_external_data(
@@ -282,7 +287,38 @@ def get_external_data(
     category,
     label,
     subdivision_type,
-):
+) -> pd.DataFrame:
+    """
+    :param session: sqlalchemy database session
+    :param jurisdiction_id: integer Id of jurisdiction in database ReportingUnit table
+    :param subdivision_type: string ReportingUnitType characterizing subdivisions for points in scatter
+    :param category: value for ExternalDataSet.Category
+    :param label: value for ExternalDataSet.Label
+    :param subdivision_type: value for ReportingUnit.ReportingUnitType
+    :return: dataframe of results rolled up to ReportingUnits of <subdivision> ReportingUnitType
+        restricted by <category> and <label>. Format is designed to match votecount format
+        Columns are:
+                "Election_Id",
+                "Name", (name of ReportingUnit)
+                "Selection", (set to <label>)
+                "Contest_Id", dummy set to 0
+                "Candidate_Id", dummy set to 0
+                "Contest", (set to <category>)
+                "CountItemType", (set to "total")
+                "Count",
+    """
+    # specify output columns
+    cols = [
+                "Election_Id",
+                "Name",
+                "Selection",
+                "Contest_Id",
+                "Candidate_Id",
+                "Contest",
+                "CountItemType",
+                "Count",
+            ]
+
     # get the census data
     census_df = db.read_external(
         session,
@@ -305,20 +341,9 @@ def get_external_data(
             columns={"Label": "Selection", "Value": "Count"},
             inplace=True,
         )
-        census_df = census_df[
-            [
-                "Election_Id",
-                "Name",
-                "Selection",
-                "Contest_Id",
-                "Candidate_Id",
-                "Contest",
-                "CountItemType",
-                "Count",
-            ]
-        ]
+        census_df = census_df[cols]
         return census_df
-    return pd.DataFrame()
+    return pd.DataFrame(columns=cols)
 
 
 def get_votecount_data(
@@ -340,7 +365,8 @@ def get_votecount_data(
[…]
