### 2021-12-20 67b6b73
Update paper.md
a few typos/suggestions
