### 2021-12-20 #3739 @ajstewartlang
Many thanks @andrewheiss and @vaneseltine for your detailed and helpful reviews - it's very much appreciated. And many thanks @sfsinger19103 for a great submission!

I've made a couple of very minor suggestions on the paper via a pull request. If you could address those @sfsinger19103 and then do the following please, that would be great:

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
