+++ src/electiondata/analyze/__init__.py
@@ -521,9 +521,9 @@ def create_bar(
 
     # package into list of dictionary
     result_list = []
-    ids = top_ranked["unit_id"].unique()
+    ids = top_ranked["bar_chart_id"].unique()
     for idx in ids:
-        temp_df = top_ranked[top_ranked["unit_id"] == idx]
+        temp_df = top_ranked[top_ranked["bar_chart_id"] == idx]
         # some cleaning here to make the pivoting work
         scores_df = temp_df[temp_df["rank"] != 1]
         scores_df = scores_df[["ReportingUnit_Id", "score", "margins_pct"]]
@@ -642,7 +642,7 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
 
     :return: dataframe obtained by appending columns to <data>:
         "score": value between  0 and 1; 1 is more anomalous
-        "unit_id": identifies the set of vote counts within which the anomaly score of the single vote count was
+        "bar_chart_id": identifies the set of vote counts within which the anomaly score of the single vote count was
             calculated. A single vote count's anomaly score depends on the set of vote counts within which it is
             considered.
         "reporting_unit_total": total votes for all selections in given contest for given reporting unit
@@ -650,7 +650,7 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
         "rank": 1 for contest winner, 2 for second place, etc.
         "contest_total": total votes cast in the contest in entire jurisdiction
         "index": integer internal id denoting the set over which the anomaly score is calculated
-        "unit_id_tmp": artifact from calculation
+        "bar_chart_id_tmp": artifact from calculation
 
     and preserving columns
         "ReportingUnit_Id" (renamed from "ParentReportingUnit_Id")
@@ -735,31 +735,31 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
         ranked_df, how="inner", on=["Contest_Id", "Selection"]
     )
 
-    # assign temporary unit_ids to unique combination of contest,
+    # assign temporary bar_chart_ids to unique combination of contest,
     # ru_type, and count type. These will be updated later to account
     # for 2 candidate pairings
     df_unit = grouped_df[
         ["Contest_Id", "ReportingUnitType", "CountItemType"]
     ].drop_duplicates()
     df_unit = df_unit.reset_index()
-    df_unit["unit_id_tmp"] = df_unit.index
+    df_unit["bar_chart_id_tmp"] = df_unit.index
     df_with_units = grouped_df.merge(
         df_unit, how="left", on=["Contest_Id", "ReportingUnitType", "CountItemType"]
     )
 
     # loop through each unit ID and assign anomaly scores
-    # also update the "real" unit_id which takes into account pairing of candidates
-    unit_ids_tmp = df_with_units["unit_id_tmp"].unique()
-    unit_id = 0 # increments on each pass through for loop
+    # also update the "real" bar_chart_id which takes into account pairing of candidates
+    bar_chart_ids_tmp = df_with_units["bar_chart_id_tmp"].unique()
+    bar_chart_id = 0 # increments on each pass through for loop
     df = pd.DataFrame() # collects records on each pass through for loop
     # for each unit ID
-    for unit_id_tmp in unit_ids_tmp:
+    for bar_chart_id_tmp in bar_chart_ids_tmp:
         # grab all the data there
-        temp_df = df_with_units[df_with_units["unit_id_tmp"] == unit_id_tmp]
+        temp_df = df_with_units[df_with_units["bar_chart_id_tmp"] == bar_chart_id_tmp]
         for i in range(2, int(temp_df["rank"].max()) + 1):
             selection_df = temp_df[temp_df["rank"].isin([1, i])].copy()
-            selection_df["unit_id"] = unit_id
[…]
