+++ src/electiondata/analyze/__init__.py
@@ -636,12 +636,12 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
         "contest_type",
         "contest_district_type",
         "Count",
-
-    and possibly other columns, such as:
-        "Party",
+        and possibly other columns, such as: "Party"
 
     :return: dataframe obtained by appending columns to <data>:
-        "score": value between  0 and 1; 1 is more anomalous
+        "score": value between 0 and 1 (1 is more anomalous) indicating, among all reporting units for the Contest,
+            how anomalous the vote share is between the Selection and the contest winner. Note that winner always
+            scores 0, as vote share of a selection with itself is always (.5,.5)
         "bar_chart_id": identifies the set of vote counts within which the anomaly score of the single vote count was
             calculated. A single vote count's anomaly score depends on the set of vote counts within which it is
             considered.
@@ -651,7 +651,6 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
         "contest_total": total votes cast in the contest in entire jurisdiction
         "index": integer internal id denoting the set over which the anomaly score is calculated
         "bar_chart_id_tmp": artifact from calculation
-
     and preserving columns
         "ReportingUnit_Id" (renamed from "ParentReportingUnit_Id")
         "Name", (renamed from "ParentName", the name of the ReportingUnit)
@@ -806,7 +805,11 @@ def assign_anomaly_score(data: pd.DataFrame) -> pd.DataFrame:
 
 def get_most_anomalous(data: pd.DataFrame, n: int) -> pd.DataFrame:
     """
-    :param data: dataframe with columns:
+    :param data: dataframe with required columns:
+        "margin_ratio": number of votes at stakedivided by overall contest margin between the two candidates
+        "score": anomaly z-score (higher is more anomalous) # TODO of chart or county?
+
+
         "ReportingUnit_Id": 
         "Name": (name of reporting unit)
         "ReportingUnitType": 
@@ -826,30 +829,30 @@ def get_most_anomalous(data: pd.DataFrame, n: int) -> pd.DataFrame:
         "bar_chart_id_tmp": artifact from calculation
         "bar_chart_id": internal integer id identifying the set of points within which the anomaly score was calculated
         "reporting_unit_total": number of votes for all candidates in the given contest and reporting unit
-        "score": anomaly z-score (higher is more anomalous)
-        "margins_pct": 
+        "margins_pct":
         "votes_at_stake": number of votes that would change if anomaly were brought in line with nearest point (see http://digitaleditions.walsworthprintgroup.com/publication/?m=7656&i=694516&p=10&ver=html5)
-       "margin_ratio": number of votes at stakedivided by overall contest margin between the two candidates
 
     :param n: integer, number of anomalous datasets to return
 
     :return:
     """
 
-    """Gets n contests, with 2 from largest votes at stake ratio
-    and 1 with largest score. If 2 from votes at stake cannot be found
+    """Gets n contests, with <n>-1 from largest votes at stake ratio
+    and 1 with largest score. If <n>-1 from votes at stake cannot be found
     (bc of threshold for score) then we fill in the top n from scores"""
-    # filter out very small votes at stake margins
+    # filter out very small votes at stake (relative to total contest margin)
     data = data[(data["margin_ratio"] > 0.01) | (data["margin_ratio"] < -0.01)]
 
-    # grab data by highest votes at stake margin (magnitude)
-    margin_data = data[data["score"] > 2.3]
-    unit_by_margin = get_unit_by_column(margin_data, "margin_ratio")
-    # grab data by highest z-score (magnitude)
[…]
