+++ src/electiondata/analyze/__init__.py
@@ -162,8 +162,8 @@ def create_scatter(
         dfv.iloc[0]["Contest"],
         jurisdiction_id,
     )
-    h_preliminary = db.is_preliminary(session,h_election_id,jurisdiction_id)
-    v_preliminary = db.is_preliminary(session,v_election_id,jurisdiction_id)
+    h_preliminary = db.is_preliminary(session, h_election_id, jurisdiction_id)
+    v_preliminary = db.is_preliminary(session, v_election_id, jurisdiction_id)
     results["preliminary"] = h_preliminary or v_preliminary
 
     # only keep the ones where there are an (x, y) to graph
@@ -181,12 +181,12 @@ def create_scatter(
 
 
 def package_results(
-        data: pd.DataFrame,
-        jurisdiction: str,
-        x: str,
-        y: str,
-        restrict: Optional[int] = None,
-) -> Dict[str,Any]:
+    data: pd.DataFrame,
+    jurisdiction: str,
+    x: str,
+    y: str,
+    restrict: Optional[int] = None,
+) -> Dict[str, Any]:
     """
     :param data: dataframe
         if "x" not equal "y", columns are "Name" (values are reporting units within the
@@ -317,15 +317,15 @@ def get_external_data(
     """
     # specify output columns
     cols = [
-                "Election_Id",
-                "Name",
-                "Selection",
-                "Contest_Id",
-                "Candidate_Id",
-                "Contest",
-                "CountItemType",
-                "Count",
-            ]
+        "Election_Id",
+        "Name",
+        "Selection",
+        "Contest_Id",
+        "Candidate_Id",
+        "Contest",
+        "CountItemType",
+        "Count",
+    ]
 
     # get the census data
     census_df = db.read_external(
@@ -464,8 +464,8 @@ def create_bar(
     :param contest_district_type: (optional string)
     :param contest_or_contest_group: (optional string) from user-facing menu, either the name of a contest or of a
             group of contests, e.g., "All congressional"
-    :param for_export: (optional)
-    :return: List of dictionaries, where each dictionary contains information to create a bar
+    :param for_export: (optional) if True, returns data for all bar charts, not just the "most interesting" ones
+    :return: Nothing (if no interesting anomalous bar charts found) ,or List of dictionaries, where each dictionary contains information to create a bar
             chart. The bar charts in the list are chosen via an algorithm favoring charts with a single outlier
             county whose impact on the margin is large. Bar charts are restricted to results for the
             <contest_or_contest_group> , if given,and also from the contests with districts of type
@@ -510,10 +510,16 @@ def create_bar(
         ranked["margins_pct"] = ranked["Count"] / ranked["reporting_unit_total"]
         ranked_margin = ranked
         votes_at_stake = calculate_votes_at_stake(ranked_margin)
-        if not for_export:
-            top_ranked = get_most_anomalous(votes_at_stake, 3)
-        else:
+        # if for export
+        if for_export:
+            # return all data
             top_ranked = votes_at_stake
+        else:
+            # otherwise return the "most interesting"
+            top_ranked = get_most_interesting(
+                votes_at_stake, constants.number_of_charts
+            )
+
     except Exception:
         return None
     if top_ranked.empty:
@@ -570,7 +576,8 @@ def create_bar(
             results = package_results(pivot_df, jurisdiction, x, y)
         else:
             results = package_results(
-                pivot_df, jurisdiction, x, y, restrict=constants.max_rus_per_bar_chart)
+                pivot_df, jurisdiction, x, y, restrict=constants.max_rus_per_bar_chart
[…]
