### 2022-01-02 #3739 @sfsinger19103
@ajstewartlang paper in `joss-submission` branch in the repo has the changes you suggested (except for that comma). It's good to go as far as I'm concerned.

Thanks again for the great feedback.
