### 2022-01-03 #3739 @sfsinger19103
Just noticed a sentence fragment in the summary. If it's not too late, how can we correct it per: https://github.com/openjournals/joss-papers/pull/2858#issuecomment-1004226020
