### 2022-01-03 #3739 @ajstewartlang
> Just noticed a sentence fragment in the summary. If it's not too late, how can we correct it per: [openjournals/joss-papers#2858 (comment)](https://github.com/openjournals/joss-papers/pull/2858#issuecomment-1004226020)

@arfon can you merge this PR please? I don't think I can.
