### 2022-01-05 #3739 @kthyng
@sfsinger19103 -

* version looks updated ✅
* Please update Zenodo metadata so that title and author list exactly match JOSS paper
* Paper comments:
- first bullet, page 2: need punctuation before "See for example"
- first paragraph under "statement of need": "Academic sources ..." should have the parenthetical references together in one set of parentheses
- references aren't properly capitalized, for example Herron 2019 and "north carolina" — preserve capitalization with {} around characters in .bib file
