### 2022-01-05 44c246d
Correct small errors in paper.md
