### 2022-01-05 5837c90
Capitalize references in paper.bib
