+++ JoSS_Submission/paper.bib
@@ -1,6 +1,6 @@
 
 @online{votevis,
-	title = {The Vote Visualizer},
+	title = {The {V}ote {V}isualizer},
 	url = {https://verifiedvoting.org/votevisualizer/},
 	author = {Singer, Stephanie Frank, Edward, Piou},
 	organization = {Verified Voting},
@@ -9,7 +9,7 @@
 	langid = {american},
 }
 @article{Herron:2019,
-	title = {Mail-In Absentee Ballot Anomalies in North Carolina's 9th Congressional District},
+	title = {Mail-In Absentee Ballot Anomalies in {N}orth {C}arolina's 9th Congressional District},
 	Volume = {18},
   	Year = 2019,
   	Month = sep,
@@ -25,7 +25,7 @@
 }
 
 @online{medsl_mit_election_data__science_lab_mit_nodate,
-	title = {MIT Election Data + Science Lab (MEDSL)},
+	title = {MIT {E}lection {D}ata + {S}cience {L}ab (MEDSL)},
 	url = {https://electionlab.mit.edu/},
 	author = {{Massachusetts Institute of Technology}},
 	year = 2021,
@@ -34,7 +34,7 @@
 
 
 @online{mcdonald_united:2021,
-	title = {United States Elections Project},
+	title = {United {S}tates Elections Project},
 	url = {http://www.electproject.org/home},
 	year = 2021,
 	author = {{McDonald}, Michael},
